`migration-tests` -> `migration-checker`
