+++ utils/migration-checker/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "migration-tests"
+name = "migration-checker"
 version = "0.1.0"
 edition = "2021"
